added a couple of monitor related api endpoints
- moved around dal methods into separate files
